Added cgroups support
