+++ pilot/util/container.py
@@ -17,7 +17,7 @@
 # under the License.
 #
 # Authors:
-# - Paul Nilsson, paul.nilsson@cern.ch, 2018-24
+# - Paul Nilsson, paul.nilsson@cern.ch, 2018-25
 
 """Functions for executing commands."""
 
@@ -39,11 +39,14 @@ from time import sleep
 from typing import Any, TextIO
 
 from pilot.common.errorcodes import ErrorCodes
+from pilot.common.pilotcache import get_pilot_cache
 #from pilot.util.loggingsupport import flush_handler
+from pilot.util.cgroups import add_process_to_cgroup
 from pilot.util.processgroups import kill_process_group
 
-logger = logging.getLogger(__name__)
 errors = ErrorCodes()
+logger = logging.getLogger(__name__)
+pilot_cache = get_pilot_cache()
 
 # Define a global lock for synchronization
 execute_lock = threading.Lock()
@@ -92,6 +95,13 @@ def execute(executable: Any, **kwargs: dict) -> Any:  # noqa: C901
                 encoding='utf-8',
                 errors='replace'
             )
+            # should we create a cgroup for the process and add the pid?
+            if pilot_cache.use_cgroups:
+                status = add_process_to_cgroup(process.pid)
+                if not status:
+                    logger.warning('failed to add process to cgroup')
+                    pilot_cache.use_cgroups = False
+
             if kwargs.get('returnproc', False):
                 return process
